add voyage_ai wrapper
This is a wrapper to bypass Voyage AI Embedding API call. 
The standard vo.embed() API call restricts usage up to 128 documents at a time. 
By using this wrapper, users can feed arbitrarily large documents.
The output can be passed on to the k_nearest_neighbours example that is provided in the Voyage API docs.